Consumer Suplier and chainign function

+++ src/LambdaIFunction.java
@@ -11,21 +11,22 @@ public class LambdaIFunction {
         function1.apply("the answer");
 
         List<Employe> employeList = new ArrayList<Employe>();
-        employeList.add(new Employe("juan",10000,0));
-        employeList.add(new Employe("kD",5000,8));
-        employeList.add(new Employe("DBook",3000,4));
-        employeList.add(new Employe("B-Beal",500,7));
+        employeList.add(new Employe("juan",10000.0,0));
+        employeList.add(new Employe("kD",5000.0,8));
+        employeList.add(new Employe("DBook",3000.0,4));
+        employeList.add(new Employe("B-Beal",500.0,7));
 
-        Function<Employe,Integer> employeIntegerFunction= employe -> {
+        Function<Employe,Double> employeIntegerFunction= employe -> {
             if(employe.salary<1000){
                 employe.salary= employe.salary *10;
             }
-            return employe.salary;
+            return (double) employe.salary;
         };
+
+
         for(Employe employee :employeList){
             employeIntegerFunction.apply(employee);
             System.out.println(employee.salary + " " + employee.name);
         }
-
     }
 }
